metatierrascol/requirements.txt
psycopg2==2.9.9
Django==4.1
djangorestframework==3.14.0
drf-yasg==1.21.7
pgOperations==0.0.3
datetime==5.3
django-filter==23.5
